Improve listing displays for Members and Family Members
